feat: added new test in serial testing sync
