+++ tests/functionnal/environnement_manager.rs
@@ -56,9 +56,9 @@ impl EnvironnementManager {
                 "run".to_string(),
                 "--bin".to_string(),
                 "service".to_string(),
-                format!("127.0.0.{new_index}:8080"),
-                format!("ws://127.0.0.{snd_index}:8080"),
-                format!("ws://127.0.0.{third_index}:8080"),
+                format!("127.0.0.{}:8081", new_index + 100),
+                format!("ws://127.0.0.{}:8081", snd_index + 100),
+                format!("ws://127.0.0.{}:8081", third_index + 100),
                 new_path.clone(),
                 new_path,
             ])
